Add type/score
