+++ packages/hello-gsm/src/type/application.ts
@@ -22,26 +22,6 @@ export interface ApplicationType {
   };
 }
 
-export interface ScoreType {
-  score2_1: number;
-  score2_2: number;
-  score3_1: number;
-  generalCurriculumScoreSubtotal: number;
-  artSportsScore: number;
-  attendanceScore: number;
-  curriculumScoreSubtotal: number;
-  volunteerScore: number;
-  nonCurriculumScoreSubtotal: number;
-  scoreTotal: number;
-  rankPercentage: number;
-}
-
-export interface GEDType {
-  curriculumScoreSubtotal: number; // 전과목 득점 합계
-  nonCurriculumScoreSubtotal: number; // 전과목 만점 합계
-  rankPercentage: number; // 백분율
-}
-
 export interface ApplyFormType {
   addressDetails: string;
   telephoneNumber: string;
